+++ 2024-2025/4.dynamic_programming/nonna_recursive.cpp
@@ -5,7 +5,6 @@ using namespace std;
 #define MAXP 1000000
 
 int N, K;
-int memo[MAXN][MAXN];
 int P[MAXN];
 
 int D(int i, int k)
@@ -16,14 +15,10 @@ int D(int i, int k)
   if (i >= N)
     return MAXP * 2;
 
-  if (!memo[i][k])
-  {
-    int non_prendo = D(i + 1, k);
-    int prendo = D(i + 1, k - P[i]) + P[i];
+  int non_prendo = D(i + 1, k);
+  int prendo = D(i + 1, k - P[i]) + P[i];
 
-    memo[i][k] = min(prendo, non_prendo);
-  }
-  return memo[i][k];
+  return min(prendo, non_prendo);
 }
 
 int main()
